add config variant_is_production
There is new loader's method, `loader.is_production()` but you have to run
`loader.setup_project_env()` beforehand.

What is considered in production?

When "variant_is_production" is truthy, or when variant is either
"prod", "production", "staging", or "stage".

We allow multiple variants to be configured in a single runner.json
file, so the usage of "variant_is_production" is not recommended.

Better if you use variant names like the ones above.